Shift to using double precision Eigen vectors and matrices (#2519)

+++ filters/ApproximateCoplanarFilter.cpp
@@ -90,7 +90,7 @@ void ApproximateCoplanarFilter::filter(PointView& view)
         auto B = eigen::computeCovariance(view, ids);
 
         // perform the eigen decomposition
-        SelfAdjointEigenSolver<Matrix3f> solver(B);
+        SelfAdjointEigenSolver<Matrix3d> solver(B);
         if (solver.info() != Success)
             throwError("Cannot perform eigen decomposition.");
         auto ev = solver.eigenvalues();
